move conversion.py from `can` to `database`
