+++ cantools/database/conversion.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
 from typing import Dict, Optional, Union
 
-from ...typechecking import Choices, SignalValueType
-from ..namedsignalvalue import NamedSignalValue
+from ..typechecking import Choices, SignalValueType
+from .namedsignalvalue import NamedSignalValue
 
 
 class BaseConversion(ABC):
